feat: refactor WaveEffect component and hooks; improve color handling and WebGL initialization
