+++ src/renderer/components/components/WaveEffect/components/WaveEffect.tsx
@@ -27,29 +27,8 @@
 import { useEffect, useRef, useState } from 'react';
 import type { ReactElement } from 'react';
 import { useWaveRenderer } from '../hooks/useWaveRenderer';
-
-/** 着色器原始背景颜色 vec3(0.002, 0.004, 0.005) */
-const SHADER_DEFAULT_BG_RGB: [number, number, number] = [0.002, 0.004, 0.005];
-
-/**
- * 将十六进制颜色转换为归一化 RGB 分量。
- * @param hex - 十六进制颜色字符串（如 #000000）。
- * @returns 归一化 RGB 三元组，各分量范围 0-1。
- */
-function hexToRgbNorm(hex: string): [number, number, number] {
-  const m = /^#?([a-f\d]{2})([a-f\d]{2})([a-f\d]{2})$/i.exec(hex);
-  if (!m) return SHADER_DEFAULT_BG_RGB;
-  return [parseInt(m[1], 16) / 255, parseInt(m[2], 16) / 255, parseInt(m[3], 16) / 255];
-}
-
-interface WaveEffectProps {
-  /** 是否播放渲染循环，默认 true。实际启动画面始终为 true，预览区按需控制。 */
-  playing?: boolean;
-  /** 背景颜色十六进制值，传入时优先于 store 读取。预览区用于实时跟随颜色选择器。 */
-  color?: string;
-  /** 波纹强调色 [r, g, b]，各分量范围 0-1。默认蓝色。 */
-  accentColor?: [number, number, number];
-}
+import { hexToRgbNorm, SHADER_DEFAULT_BG_RGB } from '../utils/color';
+import type { WaveEffectProps } from '../types';
 
 /**
  * 渲染波浪背景画布。
